cstddataplatform/requirements.txt
djangorestframework-jwt
asgiref==3.2.7
automium==0.2.6
automium-web==0.1.1
cffi==1.14.0
click==7.1.2
cryptography==2.9.2
Django==3.0.6
django-cors-headers==3.3.0
djangorestframework==3.11.0
drf-jwt==1.16.1
Flask==1.1.2
itsdangerous==1.1.0
Jinja2==2.11.2
legacy==0.1.6
MarkupSafe==1.1.1
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
simplejson==3.17.0
six==1.15.0
sqlparse==0.3.1
typing==3.7.4.1
Werkzeug==1.0.1
protobuf==3.12.2
